+++ package.json
@@ -44,6 +44,7 @@
         "postcss": "^8.4.31",
         "prettier": "^3.1.0",
         "tailwindcss": "^3.3.5",
-        "vite": "^5.0.0"
+        "vite": "^5.0.0",
+        "vite-tsconfig-paths": "^4.3.2"
     }
 }
