+++ src/components/landing/Card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import clsx from 'clsx';
-import { LandingCard } from '../../types/landingCard';
+import { LandingCard } from '@type/landingCard';
 
 export const Card: React.FC<LandingCard> = ({ title, text, link }) => (
     <div
