+++ src/components/postForm/FormGenerator.tsx
@@ -4,7 +4,7 @@ import { schema } from '../../schemas/validationSchema';
 import { FieldErrors, UseFormRegister } from 'react-hook-form';
 import InputElement from './elements/InputElement';
 import TextAreaElement from './elements/TextAreaElement';
-import PostData from '../../types/postData';
+import PostData from '../../type/postData';
 
 const FormGenerator: React.FC = () => {
     const handleSuccess = () => {
