+++ src/type/landingCard.ts
