+++ src/type/postData.ts
